Use ColorCoordinates or ActiveTCoords if none are manually mapped
With multiple textures, we can set multiple texture
coordinates, which can be mapped to each texture unit
using vtkPolyDataMapper::MapDataArrayToMultiTextureAttribute
and vtkPolyDataMapper::GetMappedTCoordsName.

If data arrays are not manually mapped to a texture unit
like described above, the texture coordinate that will be
used will be the ColorCoordinates of the mapper, or the
activeTCoords attribute of the polydata, if those exist.

We therefore also add the active TCoords to the VBO, defined
by SetTCoords/SetActiveTCoords and retrieved by GetTCoords,
and make sure than texture coordinates have a name.