Support actor texture and InterpolateScalarsBeforeMapping
With the recent changes, only textures in actor properties
were considered. The ColorTextureMap would be skipped even
if InterpolateScalarsBeforeMapping was ON, and the texture
in the vtkActor would be ignored too. This was missing in
the following loops:

1) in ReplaceShaderTCoord: the uniform for the texture unit
and the attribute array for the coordinates were not declared
and set in the shaders.
2) in SetMapperShaderParameters: the uniform for the texture
unit and the attribute array for the coordinates were not
added to the VAO.

The fix involves looping in the number of textures in the
actor property, plus 1 if InterpolateScalarsBeforeMapping
is ON, plus 1 if the actor has a texture. This allow us to
get in the different loops. After this, we need to make sure
to exit the loop when needed, since we want to ignore other
textures when interpolating scalars before mapping, or using
the actor texture.